+++ message/node-report-response_test.go
@@ -20,7 +20,7 @@ func TestNodeReportResponse(t *testing.T) {
 			Structured: message.NewNodeReportResponse(seq,
 				ie.NewNodeID("", "", "go-pfcp.epc.3gppnetwork.org"),
 				ie.NewCause(ie.CauseRequestAccepted),
-				ie.NewOffendingIE(uint16(ie.Cause)),
+				ie.NewOffendingIE(ie.Cause),
 			),
 			Serialized: []byte{
 				0x20, 0x0d, 0x00, 0x30, 0x11, 0x22, 0x33, 0x00,
