+++ message/session-establishment-response_test.go
@@ -23,7 +23,7 @@ func TestSessionEstablishmentResponse(t *testing.T) {
 				mp, fo, seid, seq, pri,
 				ie.NewNodeID("", "", "go-pfcp.epc.3gppnetwork.org"),
 				ie.NewCause(ie.CauseRequestAccepted),
-				ie.NewOffendingIE(uint16(ie.Cause)),
+				ie.NewOffendingIE(ie.Cause),
 				ie.NewFSEID(0x1111111122222222, net.ParseIP("127.0.0.1"), nil),
 				ie.NewCreatedPDR(
 					ie.NewPDRID(0xffff),
@@ -96,7 +96,7 @@ func TestSessionEstablishmentResponse(t *testing.T) {
 				mp, fo, seid, seq, pri,
 				ie.NewNodeID("", "", "go-pfcp.epc.3gppnetwork.org"),
 				ie.NewCause(ie.CauseRequestAccepted),
-				ie.NewOffendingIE(uint16(ie.Cause)),
+				ie.NewOffendingIE(ie.Cause),
 				ie.NewFSEID(0x1111111122222222, net.ParseIP("127.0.0.1"), nil),
 				ie.NewCreatedPDR(
 					ie.NewPDRID(0xffff),
